+++ package.json
@@ -4,13 +4,14 @@
   "description": "S.C.I.P.E.  Smart Computer Interface for Protocol Execution",
   "main": "src/index.js",
   "scripts": {
-  "start": "ts-node src/index.ts"
+    "start": "ts-node src/index.ts"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "discord.js": "^14.16.2",
-    "dotenv": "^16.4.5"
+    "dotenv": "^16.4.5",
+    "ts-node": "^10.9.2"
   }
 }
