+++ src/commands/misc/ping.js
@@ -1,11 +1,8 @@
-import { Client, ChatInputCommandInteraction } from "discord.js";
-
-export default {
+module.exports = {
   name: 'ping',
   description: 'Replies with the bot ping!',
-  options: [], // Ensure options are defined, even if empty
 
-  callback: async (client: Client, interaction: ChatInputCommandInteraction) => {
+  callback: async (client, interaction) => {
     await interaction.deferReply();
 
     const reply = await interaction.fetchReply();
