--- src/events/ready/consoleLog.js
@@ -1,5 +1,0 @@
-//Console logs when the bot comes online
-
-module.exports = (client) => {
-  console.log(`${client.user.tag} is online.`);
-};
